requirements.txt
numpy==1.24.3
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain_community
langchain_experimental
transformers
langchain-huggingface
-e .